template refactoring - stage 1, rename files and moving methods to correct files
