+++ open-hackathon-server/src/hackathon/hack/__init__.py
@@ -30,3 +30,4 @@ from team_manager import TeamManager
 from host_server_manager import DockerHostManager
 from azure_cert_manager import AzureCertManager
 from register_manager import RegisterManager
+from template_manager import HackathonTemplateManager
