+++ open-hackathon-server/src/hackathon/hack/team_manager.py
@@ -42,7 +42,7 @@ class TeamManager(Component):
     """Component to manage hackathon teams"""
     user_manager = RequiredFeature("user_manager")
     admin_manager = RequiredFeature("admin_manager")
-    template_manager = RequiredFeature("template_manager")
+    hackathon_template_manager = RequiredFeature("hackathon_template_manager")
 
     def get_teams_by_user(self, user_id):
         """Get all teams of specific user
@@ -315,7 +315,7 @@ class TeamManager(Component):
         """Leave a team by user
 
         :type hackathon_id: int
-        :param hid: hackathon id
+        :param hackathon_id: hackathon id
 
         :type team_name: str|unicode
         :param team_name: team name
@@ -344,7 +344,7 @@ class TeamManager(Component):
         """ team leader and admin kick some one from team
 
         :type team_name: str|unicode
-        :param tname: team name
+        :param team_name: team name
 
         :type candidate_id: int
         :param candidate_id: the candidate to kick off
@@ -410,7 +410,7 @@ class TeamManager(Component):
         if team.leader_id != g.user.id:
             return forbidden("team leader required")
         else:
-            return self.template_manager.add_template_to_hackathon(args["template_id"], team.id)
+            return self.hackathon_template_manager.add_template_to_hackathon(args["template_id"], team.id)
 
     def delete_template_from_team(self, template_id):
         """Delete template from current user's team
@@ -424,7 +424,7 @@ class TeamManager(Component):
         if team.leader_id != g.user.id:
             return forbidden("team leader required")
         else:
-            return self.template_manager.delete_template_from_hackathon(template_id, team.id)
+            return self.hackathon_template_manager.delete_template_from_hackathon(template_id, team.id)
 
     # ------------------------------ private methods ----------------------------------------
