+++ open-hackathon-server/src/hackathon/template/__init__.py
@@ -22,3 +22,6 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 """
+
+from template_library import TemplateLibrary
+from template_constants import *
