+++ open-hackathon-server/src/hackathon/template/docker_template_unit.py
@@ -23,85 +23,13 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 """
 
-import sys
-
-sys.path.append("..")
-from hackathon.template.base_template import (
-    BaseTemplate,
-)
+from template_constants import UNIT, TEMPLATE
 
 
 class DockerTemplateUnit(object):
     """
     Smallest unit in docker template
     """
-    NAME = 'name'
-    TYPE = 'type'
-    DESCRIPTION = 'description'
-    PORTS = 'ports'
-    PORTS_NAME = 'name'
-    PORTS_PORT = 'port'
-    PORTS_PUBLIC = 'public'
-    PORTS_PROTOCOL = 'protocol'
-    PORTS_URL = 'url'
-    PORTS_HOST_PORT = 'host_port'
-    PORTS_PUBLIC_PORT = 'public_port'
-    REMOTE = 'remote'
-    REMOTE_PROVIDER = 'provider'
-    REMOTE_PROTOCOL = 'protocol'
-    REMOTE_USERNAME = 'username'
-    REMOTE_PASSWORD = 'password'
-    REMOTE_PORT = 'port'
-    HOSTNAME = 'Hostname'
-    DOMAIN_NAME = 'Domainname'
-    USER = 'User'
-    ATTACH_STDIN = 'AttachStdin'
-    ATTACH_STDOUT = 'AttachStdout'
-    ATTACH_STDERR = 'AttachStderr'
-    TTY = 'Tty'
-    OPEN_STDIN = 'OpenStdin'
-    STDIN_ONCE = 'StdinOnce'
-    ENV = 'Env'
-    CMD = 'Cmd'
-    ENTRY_POINT = 'Entrypoint'
-    IMAGE = 'Image'
-    LABELS = 'Labels'
-    VOLUMES = 'Volumes'
-    WORKING_DIR = 'WorkingDir'
-    NETWORK_DISABLED = 'NetworkDisabled'
-    EXPOSED_PORTS = 'ExposedPorts'
-    MAC_ADDRESS = 'MacAddress'
-    SECURITY_OPTS = 'SecurityOpts'
-    HOST_CONFIG = 'HostConfig'
-    HOST_CONFIG_BINDS = 'Binds'
-    HOST_CONFIG_LINKS = 'Links'
-    HOST_CONFIG_LXC_CONF = 'LxcConf'
-    HOST_CONFIG_MEMORY = 'Memory'
-    HOST_CONFIG_MEMORY_SWAP = 'MemorySwap'
-    HOST_CONFIG_CPU_SHARES = 'CpuShares'
-    HOST_CONFIG_CPUSET_CPUS = 'CpusetCpus'
-    HOST_CONFIG_PORT_BINDING = 'PortBindings'
-    HOST_CONFIG_HOST_IP = 'HostIp'
-    HOST_CONFIG_HOST_PORT = 'HostPort'
-    HOST_CONFIG_PUBLISH_ALL_PORTS = 'PublishAllPorts'
-    HOST_CONFIG_PRIVILEGED = 'Privileged'
-    HOST_CONFIG_READONLY_ROOTFS = 'ReadonlyRootfs'
-    HOST_CONFIG_DNS = 'Dns'
-    HOST_CONFIG_DNS_SEARCH = 'DnsSearch'
-    HOST_CONFIG_EXTRA_HOSTS = 'ExtraHosts'
-    HOST_CONFIG_VOLUMES_FROM = 'VolumesFrom'
-    HOST_CONFIG_CAP_ADD = 'CapAdd'
-    HOST_CONFIG_CAP_DROP = 'CapDrop'
-    HOST_CONFIG_RESTART_POLICY = 'RestartPolicy'
-    HOST_CONFIG_RESTART_POLICY_NAME = 'Name'
-    HOST_CONFIG_RESTART_POLICY_MAXIMUM_RETRY_COUNT = 'MaximumRetryCount'
-    HOST_CONFIG_NETWORK_MODE = 'NetworkMode'
-    HOST_CONFIG_DEVICES = 'Devices'
-    HOST_CONFIG_ULIMITS = 'Ulimits'
-    HOST_CONFIG_LOG_CONFIG = 'LogConfig'
-    HOST_CONFIG_LOG_CONFIG_TYPE = 'Type'
-    HOST_CONFIG_LOG_CONFIG_CONFIG = 'Config'
-    HOST_CONFIG_CGROUP_PARENT = 'CgroupParent'
 
     def __init__(self, dic):
         self.dic = self.load_default_config()
@@ -110,110 +38,110 @@ class DockerTemplateUnit(object):
 
     def load_default_config(self):
         dic = {
-            self.NAME: 'web',
-            self.TYPE: 'ubuntu terminal',
-            self.DESCRIPTION: 'sample environment for ampcamp 2015',
-            self.PORTS: [
+            UNIT.NAME: 'web',
+            UNIT.TYPE: 'ubuntu terminal',
+            UNIT.DESCRIPTION: 'sample environment for ampcamp 2015',
+            UNIT.PORTS: [
                 {
-                    self.PORTS_NAME: 'Deploy',
-                    self.PORTS_PORT: 22,
-                    self.PORTS_PUBLIC: True,
-                    self.PORTS_PROTOCOL: 'tcp',
+                    UNIT.PORTS_NAME: 'Deploy',
+                    UNIT.PORTS_PORT: 22,
+                    UNIT.PORTS_PUBLIC: True,
+                    UNIT.PORTS_PROTOCOL: 'tcp',
                 }
             ],
-            self.REMOTE: {
-                self.REMOTE_PROVIDER: 'guacamole',
-                self.REMOTE_PROTOCOL: 'ssh',
-                self.REMOTE_USERNAME: 'root',
-                self.REMOTE_PASSWORD: 'root',
-                self.REMOTE_PORT: 22,
+            UNIT.REMOTE: {
+                UNIT.REMOTE_PROVIDER: 'guacamole',
+                UNIT.REMOTE_PROTOCOL: 'ssh',
+                UNIT.REMOTE_USERNAME: 'root',
+                UNIT.REMOTE_PASSWORD: 'root',
+                UNIT.REMOTE_PORT: 22,
             },
-            self.HOSTNAME: '',
-            self.DOMAIN_NAME: '',
-            self.USER: '',
-            self.ATTACH_STDIN: False,
-            self.ATTACH_STDOUT: True,
-            self.ATTACH_STDERR: True,
-            self.TTY: True,
-            self.OPEN_STDIN: True,
-            self.STDIN_ONCE: True,
-            self.ENV: [],
-            self.CMD: [],
-            self.ENTRY_POINT: '',
-            self.IMAGE: '',
-            self.LABELS: {},
-            self.VOLUMES: {},
-            self.WORKING_DIR: '',
-            self.NETWORK_DISABLED: False,
-            self.EXPOSED_PORTS: {},
-            self.MAC_ADDRESS: '',
-            self.SECURITY_OPTS: [''],
-            self.HOST_CONFIG: {
-                self.HOST_CONFIG_BINDS: [],
-                self.HOST_CONFIG_LINKS: [],
-                self.HOST_CONFIG_LXC_CONF: {},
-                self.HOST_CONFIG_MEMORY: 0,
-                self.HOST_CONFIG_MEMORY_SWAP: 0,
-                self.HOST_CONFIG_CPU_SHARES: 0,
-                self.HOST_CONFIG_CPUSET_CPUS: '',
-                self.HOST_CONFIG_PORT_BINDING: {},
-                self.HOST_CONFIG_PUBLISH_ALL_PORTS: False,
-                self.HOST_CONFIG_PRIVILEGED: False,
-                self.HOST_CONFIG_READONLY_ROOTFS: False,
-                self.HOST_CONFIG_DNS: [],
-                self.HOST_CONFIG_DNS_SEARCH: [],
-                self.HOST_CONFIG_EXTRA_HOSTS: [],
-                self.HOST_CONFIG_VOLUMES_FROM: [],
-                self.HOST_CONFIG_CAP_ADD: [],
-                self.HOST_CONFIG_CAP_DROP: [],
-                self.HOST_CONFIG_RESTART_POLICY: {
-                    self.HOST_CONFIG_RESTART_POLICY_NAME: '',
-                    self.HOST_CONFIG_RESTART_POLICY_MAXIMUM_RETRY_COUNT: 0,
+            UNIT.HOSTNAME: '',
+            UNIT.DOMAIN_NAME: '',
+            UNIT.USER: '',
+            UNIT.ATTACH_STDIN: False,
+            UNIT.ATTACH_STDOUT: True,
+            UNIT.ATTACH_STDERR: True,
+            UNIT.TTY: True,
+            UNIT.OPEN_STDIN: True,
+            UNIT.STDIN_ONCE: True,
+            UNIT.ENV: [],
+            UNIT.CMD: [],
+            UNIT.ENTRY_POINT: '',
+            UNIT.IMAGE: '',
+            UNIT.LABELS: {},
+            UNIT.VOLUMES: {},
+            UNIT.WORKING_DIR: '',
+            UNIT.NETWORK_DISABLED: False,
+            UNIT.EXPOSED_PORTS: {},
+            UNIT.MAC_ADDRESS: '',
+            UNIT.SECURITY_OPTS: [''],
+            UNIT.HOST_CONFIG: {
+                UNIT.HOST_CONFIG_BINDS: [],
+                UNIT.HOST_CONFIG_LINKS: [],
+                UNIT.HOST_CONFIG_LXC_CONF: {},
+                UNIT.HOST_CONFIG_MEMORY: 0,
+                UNIT.HOST_CONFIG_MEMORY_SWAP: 0,
+                UNIT.HOST_CONFIG_CPU_SHARES: 0,
+                UNIT.HOST_CONFIG_CPUSET_CPUS: '',
+                UNIT.HOST_CONFIG_PORT_BINDING: {},
+                UNIT.HOST_CONFIG_PUBLISH_ALL_PORTS: False,
+                UNIT.HOST_CONFIG_PRIVILEGED: False,
+                UNIT.HOST_CONFIG_READONLY_ROOTFS: False,
+                UNIT.HOST_CONFIG_DNS: [],
+                UNIT.HOST_CONFIG_DNS_SEARCH: [],
+                UNIT.HOST_CONFIG_EXTRA_HOSTS: [],
+                UNIT.HOST_CONFIG_VOLUMES_FROM: [],
+                UNIT.HOST_CONFIG_CAP_ADD: [],
+                UNIT.HOST_CONFIG_CAP_DROP: [],
+                UNIT.HOST_CONFIG_RESTART_POLICY: {
+                    UNIT.HOST_CONFIG_RESTART_POLICY_NAME: '',
+                    UNIT.HOST_CONFIG_RESTART_POLICY_MAXIMUM_RETRY_COUNT: 0,
                 },
-                self.HOST_CONFIG_NETWORK_MODE: '',
-                self.HOST_CONFIG_DEVICES: [],
-                self.HOST_CONFIG_ULIMITS: [],
-                self.HOST_CONFIG_LOG_CONFIG: {
-                    self.HOST_CONFIG_LOG_CONFIG_TYPE: 'json-file',
-                    self.HOST_CONFIG_LOG_CONFIG_CONFIG: {},
+                UNIT.HOST_CONFIG_NETWORK_MODE: '',
+                UNIT.HOST_CONFIG_DEVICES: [],
+                UNIT.HOST_CONFIG_ULIMITS: [],
+                UNIT.HOST_CONFIG_LOG_CONFIG: {
+                    UNIT.HOST_CONFIG_LOG_CONFIG_TYPE: 'json-file',
+                    UNIT.HOST_CONFIG_LOG_CONFIG_CONFIG: {},
                 },
-                self.HOST_CONFIG_CGROUP_PARENT: '',
+                UNIT.HOST_CONFIG_CGROUP_PARENT: '',
             },
         }
         return dic
 
     def set_name(self, name):
-        self.dic[self.NAME] = name
+        self.dic[UNIT.NAME] = name
 
     def get_name(self):
-        return self.dic[self.NAME]
+        return self.dic[UNIT.NAME]
 
     def get_container_config(self):
         """
         Compose post data for docker remote api create
         :return:
         """
-        for p in self.dic[self.PORTS]:
-            key = '%d/%s' % (p[self.PORTS_PORT], p[self.PORTS_PROTOCOL])
-            self.dic[self.EXPOSED_PORTS][key] = {}
-            self.dic[self.HOST_CONFIG][self.HOST_CONFIG_PORT_BINDING][key] = \
-                [{self.HOST_CONFIG_HOST_IP: '', self.HOST_CONFIG_HOST_PORT: str(p[self.PORTS_HOST_PORT])}]
-        self.dic.pop(self.NAME, "")
-        self.dic.pop(self.TYPE, "")
-        self.dic.pop(self.DESCRIPTION, "")
-        self.dic.pop(self.PORTS, None)
-        self.dic.pop(self.REMOTE, None)
-        self.dic.pop(BaseTemplate.VIRTUAL_ENVIRONMENTS_PROVIDER)
+        for p in self.dic[UNIT.PORTS]:
+            key = '%d/%s' % (p[UNIT.PORTS_PORT], p[UNIT.PORTS_PROTOCOL])
+            self.dic[UNIT.EXPOSED_PORTS][key] = {}
+            self.dic[UNIT.HOST_CONFIG][UNIT.HOST_CONFIG_PORT_BINDING][key] = \
+                [{UNIT.HOST_CONFIG_HOST_IP: '', UNIT.HOST_CONFIG_HOST_PORT: str(p[UNIT.PORTS_HOST_PORT])}]
+        self.dic.pop(UNIT.NAME, "")
+        self.dic.pop(UNIT.TYPE, "")
+        self.dic.pop(UNIT.DESCRIPTION, "")
+        self.dic.pop(UNIT.PORTS, None)
+        self.dic.pop(UNIT.REMOTE, None)
+        self.dic.pop(TEMPLATE.VIRTUAL_ENVIRONMENTS_PROVIDER)
         return self.dic
 
     def get_image_with_tag(self):
-        return self.dic[self.IMAGE]
+        return self.dic[UNIT.IMAGE]
 
     def get_ports(self):
-        return self.dic[self.PORTS]
+        return self.dic[UNIT.PORTS]
 
     def get_remote(self):
-        return self.dic[self.REMOTE]
+        return self.dic[UNIT.REMOTE]
 
     def get_image_without_tag(self):
         image = self.get_image_with_tag()
@@ -228,7 +156,7 @@ class DockerTemplateUnit(object):
             return 'latest'
 
     def get_run_command(self):
-        cmd = self.dic[self.CMD]
+        cmd = self.dic[UNIT.CMD]
         if cmd:
             return " ".join(cmd)
         return ""
@@ -241,7 +169,7 @@ class DockerTemplateUnit(object):
             if len(arr) == 2:
                 env_vars[arr[0]] = arr[1]
 
-        map(lambda env: convert(env), self.ENV or [])
+        map(lambda env: convert(env), UNIT.ENV or [])
         return env_vars
 
     def get_instance_ports(self):
@@ -255,5 +183,5 @@ class DockerTemplateUnit(object):
                 "endpoint_type": "tcp-endpoint"
             })
 
-        map(lambda p: convert(p), self.dic[self.PORTS])
+        map(lambda p: convert(p), self.dic[UNIT.PORTS])
         return instance_ports
